+++ demo/F0.Talks.AsyncAwait.NuGet/Services/NuGetService.cs
@@ -6,11 +6,11 @@ namespace F0.Talks.AsyncAwait.NuGet.Services
 {
     public static class NuGetService
     {
-        public static async Task<int> GetAsync(string packageId, bool prerelease, CancellationToken cancellationToken)
+        public static async Task<long> GetAsync(string packageId, bool prerelease, CancellationToken cancellationToken)
         {
             using var client = new NuGetClient();
 
-            int downloads = await client.GetAsync(packageId, prerelease, cancellationToken);
+            long downloads = await client.GetAsync(packageId, prerelease, cancellationToken);
 
             return downloads;
         }
